+++ CHANGELOG.md
@@ -37,6 +37,9 @@
 - Fixes Firefox compatibility issue by changing ajax to getJSON  [#263](https://github.com/openscope/openscope/issues/259)
 - Fixes bug with departures at SAME [#303](https://github.com/openscope/openscope/issues/303)
 - Fixes coordinates for PAM at EHAM [#321](https://github.com/openscope/openscope/issues/321)
+- Ensure aircraft reach their targeted speed [#340](https://github.com/openscope/openscope/issues/340)
+- Fixes last-second go-arounds by landing aircraft [#342](https://github.com/openscope/openscope/issues/342)
+- Ensure aircraft follow glideslope [#346](https://github.com/openscope/openscope/issues/346)
 
 
 ## 4.1.2 (February 20, 2017)